requirements.txt
Adafruit-Blinka==8.62.0
adafruit-circuitpython-busdevice==5.2.13
adafruit-circuitpython-connectionmanager==3.1.5
adafruit-circuitpython-sht4x==1.0.24
adafruit-circuitpython-dht==4.0.9
adafruit-circuitpython-requests==4.1.13
adafruit-circuitpython-typing==1.12.1
Adafruit-PlatformDetect==3.81.0
Adafruit-PureIO==1.1.11
annotated-types==0.7.0
anyio==4.9.0
binho-host-adapter==0.1.6
certifi==2025.7.14
charset-normalizer==3.4.2
click==8.2.1
dnspython==2.7.0
email_validator==2.2.0
fastapi==0.116.1
fastapi-cli==0.0.8
fastapi-cloud-cli==0.1.4
h11==0.16.0
httpcore==1.0.9
httptools==0.6.4
httpx==0.28.1
idna==3.10
Jinja2==3.1.6
markdown-it-py==3.0.0
MarkupSafe==3.0.2
mdurl==0.1.2
pydantic==2.11.7
pydantic_core==2.33.2
pyftdi==0.56.0
Pygments==2.19.2
pyserial==3.5
python-dotenv==1.1.1
python-multipart==0.0.20
pyusb==1.3.1
PyYAML==6.0.2
requests==2.32.4
rich==14.0.0
rich-toolkit==0.14.8
rignore==0.6.4
RPi.GPIO==0.7.1
sentry-sdk==2.33.1
shellingham==1.5.4
sniffio==1.3.1
spidev==3.7
starlette==0.47.2
sysv_ipc==1.1.0
typer==0.16.0
typing-inspection==0.4.1
typing_extensions==4.14.1
urllib3==2.5.0
uvicorn==0.35.0
uvloop==0.21.0
watchfiles==1.1.0
websockets==15.0.1